Number, Shape enum을 Card의 nested enum으로 변경
