--- PokerGameApp/PokerGameApp/Shape.swift
@@ -1,15 +1,0 @@
-//
-//  Shape.swift
-//  PokerGameApp
-//
-//  Created by Bibi on 2022/02/23.
-//
-
-import Foundation
-
-enum Shape: String {
-    case spades = "♠︎"
-    case hearts = "♥︎"
-    case clubs = "♣︎"
-    case diamonds = "♦︎"
-}
